Fixing records deserialization issue (#5094, parts of #4628) with read-access when json renamed to same property name (#5073)
