+++ release-notes/CREDITS-2.x
@@ -1885,3 +1885,8 @@ Zhen Lin Low (@zhenlin-pay2)
  * Reported, fixed #4920: Creator properties are ignored on abstract types
    when collecting bean properties, breaking AsExternalTypeDeserializer
   (2.18.3)
+
+Fawzi Essam (@iifawz)
+ * Contributed fix for #5049: Duplicate creator property "b" (index 0 vs 1)
+   on simple java record
+  (2.18.3)
