+++ src/test-jdk17/java/com/fasterxml/jackson/databind/records/RecordWithReadOnlyTest.java
@@ -7,6 +7,8 @@ import com.fasterxml.jackson.databind.testutil.DatabindTestUtil;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
 
 public class RecordWithReadOnlyTest extends DatabindTestUtil {
 
@@ -50,6 +52,27 @@ public class RecordWithReadOnlyTest extends DatabindTestUtil {
         }
     }
 
+
+    static class ReadOnly5049Pojo
+    {
+        protected String a, b;
+
+        ReadOnly5049Pojo(
+                @JsonProperty(value = "a", access = JsonProperty.Access.READ_ONLY) String a,
+                @JsonProperty(value = "b", access = JsonProperty.Access.READ_ONLY) String b) {
+            this.a = a;
+            this.b = b;
+        }
+
+        public String getA() { return a; }
+        public String getB() { return b; }
+    }
+
+    record ReadOnly5049Record(
+            @JsonProperty(value = "a", access = JsonProperty.Access.READ_ONLY) String a,
+            @JsonProperty(value = "b", access = JsonProperty.Access.READ_ONLY) String b) {
+    }
+
     private final ObjectMapper MAPPER = newJsonMapper();
 
     /*
@@ -82,14 +105,31 @@ public class RecordWithReadOnlyTest extends DatabindTestUtil {
         assertEquals(a2q("{'id':123,'name':'Bob'}"), json);
     }
 
-    /**
-     * Currently documents a bug where a property was NOT ignored during deserialization if given an explicit name.
-     * Also reproducible in 2.14.x.
-     */
+
     @Test
     public void testDeserializeReadOnlyNamedProperty() throws Exception {
         RecordWithReadOnlyNamedProperty value = MAPPER.readValue(a2q("{'id':123,'name':'Bob'}"), RecordWithReadOnlyNamedProperty.class);
-        assertEquals(new RecordWithReadOnlyNamedProperty(123, "Bob"), value); // BUG: should be `null` instead of "Bob"
+        assertEquals(new RecordWithReadOnlyNamedProperty(123, null), value);
+    }
+
+    @Test
+    void testRoundtripPOJO() throws Exception
+    {
+        String json = MAPPER.writeValueAsString(new ReadOnly5049Pojo("hello", "world"));
+        ReadOnly5049Pojo pojo = MAPPER.readerFor(ReadOnly5049Pojo.class).readValue(json);
+        assertNotNull(pojo);
+        assertNull(pojo.a);
+        assertNull(pojo.b);
+    }
+
+    @Test
+    void testRoundtripRecord() throws Exception
+    {
+        String json = MAPPER.writeValueAsString(new ReadOnly5049Record("hello", "world"));
+        ReadOnly5049Record record = MAPPER.readValue(json, ReadOnly5049Record.class);
+        assertNotNull(record);
+        assertNull(record.a());
+        assertNull(record.b());
     }
 
     /*
